+++ certbot_zimbra.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
 # author: Lorenzo Milesi <maxxer@yetopen.it>
-# contributions: Jernej Jakob <jernej.jakob@gmail.com>
+# author: Jernej Jakob <jernej.jakob@gmail.com>
 # GPLv3 license
 
 PROGNAME="certbot-zimbra"
-VERSION="0.7.7"
+VERSION="0.7.11"
 GITHUB_URL="https://github.com/YetOpen/certbot-zimbra"
 # paths
 ZMPATH="/opt/zimbra"
@@ -14,6 +14,9 @@ LE_LIVE_PATH="/etc/letsencrypt/live" # the domain will be appended to this path
 TEMPPATH="/run/$PROGNAME"
 # other options
 ZMPROV_OPTS="-l" # use ldap (faster)
+# used to extract the CA for the letsencrypt certs
+ca_certificates_file="/etc/ssl/certs/ca-certificates.crt"
+pki_ca_bundle_file="/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem"
 # Do NOT modify anything after this line.
 WEBROOT=""
 CERTPATH=""
@@ -32,6 +35,8 @@ DETECT_PUBLIC_HOSTNAMES=true
 SKIP_PORT_CHECK=false
 PORT=""
 QUIET=false
+MIN_CERTBOT_VERSION="0.19.0"
+LOCKED=false
 
 # set up a trap on exit
 exitfunc(){
@@ -41,6 +46,13 @@ exitfunc(){
 		echo "An error seems to have occurred. Please read the output above for clues and try to rectify the situation."
 		echo "If you believe this is an error with the script, please file an issue at $GITHUB_URL."
 	fi
+
+	# close fd used for locking
+	exec 200>&-
+	if "$LOCKED"; then
+		rm "$TEMPPATH/$PROGNAME.lck"
+	fi
+
 	exit "$e"
 }
 trap exitfunc EXIT
@@ -61,7 +73,9 @@ make_temp() {
 
 get_lock(){
 	exec 200> "$TEMPPATH/$PROGNAME.lck"
-	! flock -n 200 && echo "Error: can't get exclusive lock. Another instance of this script may be running." && exit 1
+	! flock -n 200 && echo "Error: can't get exclusive lock. Another instance of this script may be running.
+If you are sure there is no other instance of this script running (check with \"ps afx\") you can remove $TEMPPATH/$PROGNAME.lck and try again." && exit 1
+	LOCKED=true
 }
 
 prompt(){
@@ -76,11 +90,11 @@ prompt(){
 
 check_depends_ca() {
 	# Debian/Ubuntu provided by ca-certificates
-	[ -f /etc/ssl/certs/ca-certificates.crt ] && return
+	[ -r $ca_certificates_file ] && return
 	# RHEL/CentOS provided by pki-base
-	[ -f /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem ] && return
+	[ -r $pki_ca_bundle_file ] && return
 
-	echo "Installed CA certificates not found. Please check if you have installed:"
+	echo "Error: Installed CA certificates not found or files not readable. Please check if you have installed:"
 	echo "Debian/Ubuntu: ca-certificates (if you do, you might have to run \"update-ca-certificates\")"
 	echo "RHEL/CentOS: pki-base (if you do, you might have to run \"update-ca-trust\")"
 	exit 1
@@ -91,7 +105,7 @@ check_depends() {
 	! $QUIET && echo "Checking for dependencies..."
 
 	# do not check for lsof or ss here as we'll do that later
-	for name in su openssl grep head cut sed chmod chown cat cp $ZMPATH/bin/zmcertmgr $ZMPATH/bin/zmcontrol $ZMPATH/bin/zmprov $ZMPATH/libexec/get_plat_tag.sh; do
+	for name in su openssl grep head cut sed chmod chown cat cp gawk $ZMPATH/bin/zmhostname $ZMPATH/bin/zmcertmgr $ZMPATH/bin/zmcontrol $ZMPATH/bin/zmprov $ZMPATH/libexec/get_plat_tag.sh; do
 		! which "$name" >/dev/null && echo "Error: \"$name\" not found or executable" && exit 1
 	done
 }
@@ -117,6 +131,8 @@ bootstrap() {
 	[ -z "$DETECTED_ZIMBRA_VERSION" ] && echo "Error: Unable to detect zimbra version" && exit 1
 	! "$QUIET" && echo "Detected Zimbra $DETECTED_ZIMBRA_VERSION on $PLATFORM"
 
+	get_domain
+
 	return 0
 }
 
@@ -126,12 +142,10 @@ check_zimbra_proxy() {
 
 	! "$QUIET" && echo "Checking zimbra-proxy is running and enabled"
 
-	# no need if we check if it's running later
-	#su - zimbra -c "$ZMPATH/bin/zmprov $ZMPROV_OPTS gs $DOMAIN zimbraServiceEnabled | grep -q proxy" || ( echo "Error: zimbra-proxy is not enabled" && exit 1 )
-
 	# TODO: check if path to zmproxyctl is different on <8.7
 	! su - zimbra -c "$ZMPATH/bin/zmproxyctl status > /dev/null" && echo "Error: zimbra-proxy is not running" && exit 1
-	! su - zimbra -c "$ZMPATH/bin/zmprov $ZMPROV_OPTS gs $DOMAIN zimbraReverseProxyHttpEnabled | grep -q TRUE" && echo "Error: http reverse proxy not enabled (zimbraReverseProxyHttpEnabled: FALSE)" && exit 1
+	! su - zimbra -c "$ZMPATH/bin/zmprov $ZMPROV_OPTS gs $DOMAIN zimbraReverseProxyHttpEnabled | grep -q TRUE" \
+			&& echo "Error: http reverse proxy not enabled (zimbraReverseProxyHttpEnabled: FALSE)" && exit 1
 
 	if [ -z "$PORT" ]; then
 		! "$QUIET" && echo "Detecting port from zimbraMailProxyPort"
@@ -139,6 +153,19 @@ check_zimbra_proxy() {
 		[ -z "$PORT" ] && echo "Error: zimbraMailProxyPort not found" && exit 1
 	else
 		echo "Skipping port detection from zimbraMailProxyPort due to --port override"
+	fi
+
+	if [ "$PORT" != "80" ] && ! "$QUIET"; then
+		echo "WARNING: non-standard zimbraMailProxyPort $PORT. 
+This needs to be 80 from the internet for Let's Encrypt (certbot) to work. 
+If you have NAT set up to do the translation this is likely fine. 
+If not, your Zimbra proxy is misconfigured and certbot will fail."
+		if "$PROMPT_CONFIRM"; then
+			prompt "Proceed?"
+			(( $? == 1 )) && echo "Cannot continue. Exiting." && exit 0
+		else
+			echo "WARNING: Prompt disabled, proceeding anyway."
+		fi
 	fi
 }
 
@@ -150,17 +177,21 @@ check_port () {
 	fi
 
 	[ -z "$1" ] && echo 'Unexpected error: check_port empty $1 (port)' && exit 1
-	
+
 	! "$QUIET" && echo "Checking if process is listening on port $1 ${2:+"with name \"$2\" "}${3:+"user \"$3\""}"
 
 	# check with lsof if available, or fall back to ss
 	LSOF_BIN="$(which lsof 2>/dev/null)"
 	SS_BIN="$(which ss 2>/dev/null)"
 	CHECK_BIN=""
+	GREP_FILTER=""
 	if [ -x "$LSOF_BIN" ]; then
 		CHECK_BIN="$LSOF_BIN -i :$1 -s TCP:LISTEN -a -n"
+		GREP_FILTER="$2.*$3"
 	elif [ -x "$SS_BIN" ]; then
 		CHECK_BIN="$SS_BIN -lptn sport eq :$1"
+		# ss doesn't return process user, so don't use it for count
+		GREP_FILTER="$2"
 	else
 		echo 'Error: Neither "lsof" nor "ss" were found in PATH. Unable to continue, exiting.'
 		exit 1
@@ -168,7 +199,7 @@ check_port () {
 
 	# return false if count of matched processes is 0
 	# optionally $2 and $3 are process name and process user
-	(( "$($CHECK_BIN | grep -c "$2.*$3")" == 0 )) && return 1
+	(( "$($CHECK_BIN | grep -c "$GREP_FILTER")" == 0 )) && return 1
 
 	return 0
 }
@@ -178,44 +209,86 @@ check_port () {
 # zimbra-proxy must be running (checked with check_zimbra_proxy) or zmproxyctl restart will fail
 # returns true if patch was applied or was already present, exits script if encountered an error
 patch_nginx() {
-	[ ! -d $ZMPATH/conf/nginx/includes ] && echo "Error: $ZMPATH/conf/nginx/includes not found, exiting" && exit 1
+	[ ! -d "$ZMPATH/conf/nginx/includes" ] && echo "Error: $ZMPATH/conf/nginx/includes not found, exiting" && exit 1
 
-	# Return if patch is already applied
-	if grep -r -q 'acme-challenge' "$ZMPATH/conf/nginx/includes"; then
+	# Don't patch if patch is already applied
+	if grep -r -q 'acme-challenge' "$ZMPATH/conf/nginx/templates"; then
 		! "$QUIET" && echo "Nginx templates already patched."
-		return
+	else
+		[ -z $WEBROOT ] && echo "Unexpected error: patch_nginx WEBROOT not set. Exiting." && exit 1
+
+		# Let's make a backup of zimbra's original templates
+		set -e
+		BKDATE="$(date +'%Y%m%d_%H%M%S')"
+		! "$QUIET" && echo "Making a backup of nginx templates in $ZMPATH/conf/nginx/templates.$BKDATE"
+		cp -a "$ZMPATH/conf/nginx/templates" "$ZMPATH/conf/nginx/templates.$BKDATE"
+		set +e
+
+		# do patch
+		! "$QUIET" && echo -n "Patching nginx templates... "
+		e=0
+		for file in http.default https.default http https ; do
+			# Find the } that matches the first { after the server directive and add our location block before it,
+			# ignoring all curly braces and anything else between them. If there are multiple server blocks
+			# it adds the directives to all of them. It breaks in special cases of one-liner server blocks (rare)
+			# and unbalanced curly brace count (missing braces aka broken formatting).
+			# Exits 0 (success) if at least 1 substitution was made, 1 (failure) if 0 substitutions were made.
+			gawk \
+"BEGIN {e = 1}
+/^#/ {print; next}
+/^server[[:space:]{]*.*$/ {found++}
+/{/ && found {
+  b++
+  if (first == 0) first = NR
+}
+/}/ && found {b--}
+{ if (found && b == 0 && first != 0) {
+    print gensub(/}[^}]*/, \"\n    # patched by certbot-zimbra.sh\n    location ^~ /.well-known/acme-challenge {\n        root $WEBROOT;\n    }\n&\", 1)
+    found = 0
+    first = 0
+    e = 0
+  }
+  else print
+}
+END {exit e}" "$ZMPATH/conf/nginx/templates.$BKDATE/nginx.conf.web.$file.template" > "$ZMPATH/conf/nginx/templates/nginx.conf.web.$file.template"
+			e="$?"
+			[ "$e" -ne 0 ] && break
+		done
+
+		if [ "$e" -ne 0 ]; then
+			! "$QUIET" && echo -ne "Error!\nRestoring old templates... "
+			cp -a $ZMPATH/conf/nginx/templates.$BKDATE/* "$ZMPATH/conf/nginx/templates/"
+			if [ "$?" -ne 0 ]; then
+				! "$QUIET" && echo "Error!"
+			else
+				! "$QUIET" && echo "Success."
+			fi
+			! "$QUIET" && echo "Exiting." && exit 1
+		else
+			! "$QUIET" && echo "Success."
+		fi
 	fi
 
-	[ -z $WEBROOT ] && echo "Unexpected error: patch_nginx WEbROOT not set. Exiting." && exit 1
+	# Don't restart if includes show nginx has already been restarted
+	if grep -r -q 'acme-challenge' "$ZMPATH/conf/nginx/includes"; then
+		! "$QUIET" && echo "Nginx includes already patched, skipping zmproxy restart."
+	else
+		if "$PROMPT_CONFIRM"; then
+			prompt "Restart zmproxy?"
+			(( $? == 1 )) && echo "Cannot continue. Exiting." && exit 0
+		fi
 
-	! "$QUIET" && echo "Patching nginx templates."
-	
-	set -e
-
-	# Let's make a backup of zimbra's original templates
-	BKDATE="$(date +'%Y%m%d_%H%M%S')"
-	! "$QUIET" && echo "Making a backup of nginx templates in $ZMPATH/conf/nginx/templates.$BKDATE"
-	cp -a "$ZMPATH/conf/nginx/templates" "$ZMPATH/conf/nginx/templates.$BKDATE"
-
-	# do patch
-	for file in http.default https.default http https ; do
-		sed -i "s#^}#\n    \# patched by certbot-zimbra.sh\n    location ^~ /.well-known/acme-challenge {\n        root $WEBROOT;\n    }\n}#" "$ZMPATH/conf/nginx/templates/nginx.conf.web.$file.template"
-	done
-
-	set +e
-
-	if ! "$QUIET" && "$PROMPT_CONFIRM"; then
-                prompt "Restart zmproxy?"
-                (( $? == 1 )) && echo "Cannot continue. Exiting." && exit 1
-        fi
-
-	! "$QUIET" && echo "Running zmproxyctl restart."
-	# reload nginx config
-	su - zimbra -c 'zmproxyctl restart'; e="$?"
-	if [ "$e" -ne 0 ]; then
-		echo "Error restarting zmproxy (zmproxyctl exit status $e). Exiting."
-		exit 1
+		! "$QUIET" && echo "Running zmproxyctl restart."
+		# reload nginx config
+		su - zimbra -c 'zmproxyctl restart' 200>&-; e="$?"
+		if [ "$e" -ne 0 ]; then
+			! "$QUIET" && echo "Error restarting zmproxy (\"zmproxyctl restart\" exit status $e). Exiting."
+			exit 1
+		else
+			! "$QUIET" && echo "Success."
+		fi
 	fi
+
 	return 0
 }
 
@@ -230,16 +303,12 @@ find_additional_public_hostnames() {
 		return
 	fi
 
-	! "$QUIET" && echo -n "Detecting additional public service hostnames..."
-	for i in $($ZMPATH/bin/zmprov $ZMPROV_OPTS gad); do
-		getdomain="$($ZMPATH/bin/zmprov $ZMPROV_OPTS gd $i zimbraPublicServiceHostname | grep zimbraPublicServiceHostname | cut -f 2 -d ' ')"
-		[ -z "$getdomain" ] && continue
-		# Skip our primary domain
-		[ "$getdomain" == "$DOMAIN" ] && continue
-		EXTRA_DOMAINS=("${EXTRA_DOMAINS[@]}" "$getdomain")
-		! "$QUIET" && echo -n " $getdomain"
-	done
-	! "$QUIET" && echo
+	! "$QUIET" && echo -n "Detecting additional public service hostnames... "
+
+	EXTRA_DOMAINS=($(su - zimbra -c "zmprov $ZMPROV_OPTS gad" | gawk '{printf "gd %s zimbraPublicServiceHostname\n", $0}' \
+			| su - zimbra -c "zmprov $ZMPROV_OPTS -" | sed "/prov>/d;/# name/d;/$DOMAIN/d;/^$/d;s/zimbraPublicServiceHostname: \(.*\)/\1/"))
+	! "$QUIET" && echo "Found ${#EXTRA_DOMAINS[@]} zimbraPublicServiceHostnames through auto-detection"
+
 	return 0
 }
 
@@ -248,19 +317,28 @@ get_domain () {
 	if [ -z "$DOMAIN" ]; then
 		! "$QUIET" && echo "Using zmhostname to detect domain."
 		DOMAIN="$($ZMPATH/bin/zmhostname)"
-		# Find additional domains
-		"$DEPLOY_ONLY" || find_additional_public_hostnames
 	fi
 
 	[ -z "$DOMAIN" ] && echo "Error: No domain found! Please run with -d/--hostname or check why zmhostname is not working" && exit 1
-	
-	! "$QUIET" && echo "Using domain $DOMAIN (as certificate DN)"
-	[ -n "$EXTRA_DOMAINS" ] && ! "$QUIET" && echo "Found domains to use as certificate SANs: ${EXTRA_DOMAINS[@]}"
 
-	if ! "$QUIET" && "$PROMPT_CONFIRM"; then
+	! "$QUIET" && echo "Using domain $DOMAIN (as certificate DN)"
+
+	if "$PROMPT_CONFIRM"; then
 		prompt "Is this correct?"
-		(( $? == 1 )) && echo "Error: Please call $(basename $0) --hostname your.host.name" && exit 1
+		(( $? == 1 )) && echo "Error: Please call $(basename $0) --hostname your.host.name" && exit 0
 	fi
+
+	# Find additional domains
+	"$NEW_CERT" && find_additional_public_hostnames
+
+	if [ -n "$EXTRA_DOMAINS" ] && ! "$QUIET"; then
+		echo "Got ${#EXTRA_DOMAINS[@]} domains to use as certificate SANs: ${EXTRA_DOMAINS[@]}"
+		if "$PROMPT_CONFIRM"; then
+			prompt "Include these in the certificate?"
+			(( $? == 1 )) && unset EXTRA_DOMAINS
+		fi
+	fi
+
 	return 0
 }
 
@@ -290,21 +368,17 @@ set_certpath() {
 
 check_webroot () {
 	[ -z "$WEBROOT" ] && echo "Unexpected error: check_webroot WEBROOT not set. Exiting." && exit 1
-		
+	
 	# <8.7 didn't have nginx webroot
 	if [ ! -d "$WEBROOT" ]; then
-		if ! "$QUIET" && "$PROMPT_CONFIRM"; then
-			prompt "Create $WEBROOT?"
-			if (( $? == 0 )); then
-				mkdir -p "$WEBROOT"
-				return 0
-			else
-				echo "Cannot proceed, exiting."
-				exit 1
-			fi
+		if "$PROMPT_CONFIRM"; then
+			prompt "Webroot $WEBROOT doesn't exist, create it?"
+			(( $? == 1 )) && echo "Cannot proceed, exiting." && exit 0
 		fi
-		echo "Error: $WEBROOT does not exist, cannot proceed. Please create it manually or rerun this script with -c/--prompt-confirm and without -q/--quiet. Exiting."
-		exit 1
+		echo "Creating webroot $WEBROOT"
+		set -e
+		mkdir -p "$WEBROOT"
+		set +e
 	fi
 }
 
@@ -312,20 +386,56 @@ find_certbot () {
 	# check for executable certbot-auto / certbot / letsencrypt
 	LE_BIN="$(which certbot-auto certbot letsencrypt 2>/dev/null | head -n 1)"
 	[ -z "$LE_BIN" ] && echo "Error: No letsencrypt/certbot binary found in $PATH" && exit 1
+
+	! "$QUIET" && echo "Detecting certbot version..."
+
+	# get certbot version, we need to use some trickery here in case certbot is bootstrapping (1st run) and expects user input
+	# if run with --prompt-confirm, show the output of certbot on stderr and allow the user to answer yes/no
+	# otherwise add --no-bootstrap and exit in case of error
+	local certbot_version_params=""
+	! "$PROMPT_CONFIRM" && certbot_version_params="--no-bootstrap "
+	"$LE_NONIACT" && certbot_version_params+="--non-interactive "
+	certbot_version_params+="--version"
+
+	local detected_certbot_version="$($LE_BIN ${certbot_version_params} 2>&1 | $( "$PROMPT_CONFIRM" && echo 'tee /dev/stderr |') grep '^certbot .*$')"
+	certbot_version_exit=${PIPESTATUS[0]}
+	if [ "$certbot_version_exit" -ne 0 ]; then
+		! "$QUIET" && echo "Error: \"$LE_BIN ${certbot_version_params}\" exit status $certbot_version_exit.
+Try running $LE_BIN by itself on the command line and see if it works (it may need to bootstrap itself).
+Exiting."
+		exit 1
+	fi
+
+	if [ -z "$detected_certbot_version" ]; then
+		! "$QUIET" && echo "Error: unable to parse certbot version. Exiting."
+		exit 1
+	fi
+
+	! "$QUIET" && ! "$PROMPT_CONFIRM" && echo "Detected $detected_certbot_version"
+
+	if ! version_gt "$(echo "$detected_certbot_version" | grep -Po '(\d+).(\d+).(\d+)')" "$MIN_CERTBOT_VERSION"; then
+		! "$QUIET" && echo "Error: certbot is too old, please upgrade to certbot >=$MIN_CERTBOT_VERSION. Exiting."
+		exit 1
+	fi
+
 	return 0
 }
 
 # perform the letsencrypt request
 request_cert() {
 	check_webroot
-	
+
+	if "$PROMPT_CONFIRM"; then
+		prompt "We will now run certbot to request the certificate. Proceed?"
+		(( $? == 1 )) && echo "Exiting." && exit 0
+	fi
+
 	#TODO: dry-run
-	
-	"$LE_NONIACT" && LE_PARAMS="--non-interactive"
+
+	"$LE_NONIACT" && LE_PARAMS="$LE_PARAMS --non-interactive"
 	"$QUIET" && LE_PARAMS="$LE_PARAMS --quiet"
 	"$LE_AGREE_TOS" && LE_PARAMS="$LE_PARAMS --agree-tos"
-	# use --cert-name instead of --expand as it allows also removing domains? https://github.com/certbot/certbot/issues/4275
-	LE_PARAMS="$LE_PARAMS --webroot -w $WEBROOT --expand -d $DOMAIN"
+	LE_PARAMS="$LE_PARAMS --webroot -w $WEBROOT --cert-name $DOMAIN -d $DOMAIN"
 	for d in ${EXTRA_DOMAINS[@]}; do
 		[ -z "$d" ] && continue
 		LE_PARAMS="$LE_PARAMS -d $d"
@@ -333,10 +443,12 @@ request_cert() {
 
 	! "$QUIET" && echo "Running $LE_BIN certonly $LE_PARAMS"
 	"$QUIET" && exec > /dev/null
+	"$QUIET" && exec 2>/dev/null
 	# Request our cert
-	$LE_BIN certonly $LE_PARAMS
+	"$LE_BIN" certonly $LE_PARAMS
 	e=$?
-	"$QUIET" && exec > /dev/tty
+	"$QUIET" && exec > /dev/stdout
+	"$QUIET" && exec 2> /dev/stderr
 	[ "$e" -ne 0 ] && echo "Error: $LE_BIN exit status $e. Cannot proceed, exiting." && exit 1
 	return 0
 }
@@ -347,7 +459,7 @@ prepare_cert() {
 
 	[ -z "$CERTPATH" ] && echo "Unexpected error (prepare_cert CERTPATH not set). Exiting." && exit 1
 	[ -z "$DOMAIN" ] && echo "Unexpected error (prepare_cert DOMAIN not set). Exiting." && exit 1
-	
+
 	# Make zimbra accessible files
 	# save old umask
 	oldumask="$(umask -p)"
@@ -363,20 +475,23 @@ prepare_cert() {
 
 	# Create the "patched" chain suitable for Zimbra
 	cat "$CERTPATH/chain.pem" > "$tmpcerts/zimbra_chain.pem"
-	if [ -f /etc/ssl/certs/ca-certificates.crt ]; then
+	if [ -r "$ca_certificates_file" ]; then
 	        # Debian/Ubuntu
 		# use the issuer_hash of the LE chain cert to find the root CA in /etc/ssl/certs
-		cat "/etc/ssl/certs/$(openssl x509 -in $CERTPATH/chain.pem -noout -issuer_hash).0" >> $tmpcerts/zimbra_chain.pem
-	elif [ -f /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem ]; then
+		cat "/etc/ssl/certs/$(openssl x509 -in $CERTPATH/chain.pem -noout -issuer_hash).0" >> "$tmpcerts/zimbra_chain.pem"
+	elif [ -r "$pki_ca_bundle_file" ]; then
 		# RHEL/CentOS
 		# extract CA by CN in tls-ca-bundle.pem
-		issuer="$(openssl x509 -in $CERTPATH/chain.pem -noout -issuer | sed -n 's/.*CN=//;s/\/.*$//;p')"
-		[ -z "$issuer" ] && exit 1
+		issuer="$(openssl x509 -in $CERTPATH/chain.pem -noout -issuer | sed -n 's/.*CN\s\?=\s\?//;s/\/.*$//;p')"
+		[ -z "$issuer" ] && echo "Error: can't find issuer of \"$CERTPATH/chain.pem\". Exiting." && exit 1
+		# if we can't find the issuer in the bundle file, it may have spaces removed, hopefully we'll find it without spaces
+		grep -q "^# $issuer\$" "$pki_ca_bundle_file" || issuer="${issuer//' '}"
 		# the following awk script extracts the CA cert from the bundle or exits 1 if not found
-		awk "BEGIN {e=1}; /^# $issuer$/{e=0} /^# $issuer$/,/END CERTIFICATE/; END {exit e}" /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem >> $tmpcerts/zimbra_chain.pem
+		! gawk "BEGIN {e=1}; /^# $issuer$/{e=0} /^# $issuer$/,/END CERTIFICATE/; END {exit e}" "$pki_ca_bundle_file" >> "$tmpcerts/zimbra_chain.pem"\
+			&& echo "Error: Can't find \"$issuer\" in \"$pki_ca_bundle_file\". Exiting." && exit 1
 	else
 		# we shouldn't be here
-		echo "Unexpected error (problem in check_depends_ca)" && exit 1
+		echo "Error in prepare_cert: can't find installed CA certificates (check_depends_ca should have caught this). Exiting." && exit 1
 	fi
 
 	$oldumask
@@ -390,16 +505,18 @@ prepare_cert() {
 
 	# redirect stdout to /dev/null if quiet
 	"$QUIET" && exec > /dev/null
+	"$QUIET" && exec 2>/dev/null
 
 	# Test cert. 8.6 and below must use root
 	if version_gt "$DETECTED_ZIMBRA_VERSION" "8.7"; then
 		su - zimbra -c "$ZMPATH/bin/zmcertmgr verifycrt comm $tmpcerts/privkey.pem $tmpcerts/cert.pem $tmpcerts/zimbra_chain.pem"
 	else
-		$ZMPATH/bin/zmcertmgr verifycrt comm "$tmpcerts/privkey.pem" "$tmpcerts/cert.pem" "$tmpcerts/zimbra_chain.pem"
+		"$ZMPATH/bin/zmcertmgr" verifycrt comm "$tmpcerts/privkey.pem" "$tmpcerts/cert.pem" "$tmpcerts/zimbra_chain.pem"
 	fi
 
 	# undo quiet
-	"$QUIET" && exec > /dev/tty
+	"$QUIET" && exec > /dev/stdout
+	"$QUIET" && exec 2> /dev/stderr
 
 	# undo set -e
 	set +e
@@ -409,6 +526,11 @@ prepare_cert() {
 
 # deploys certificate and restarts zimbra. ASSUMES prepare_certificate has been called already
 deploy_cert() {
+	if "$PROMPT_CONFIRM"; then
+		prompt "Deploy certificates to Zimbra? This may restart some services."
+		(( $? == 1 )) && echo "Cannot proceed. Exiting." && exit 0
+	fi
+
 	# exit on error
 	set -e
 	! "$QUIET" && echo "Deploying certificates."
@@ -418,41 +540,40 @@ deploy_cert() {
 
 	# copy privkey
 	cp -a "$tmpcerts/privkey.pem" "$ZMPATH/ssl/zimbra/commercial/commercial.key"
-	
-	if ! "$QUIET" && "$PROMPT_CONFIRM"; then
-		prompt "Deploy certificates to Zimbra?"
-		(( $? == 1 )) && echo "Cannot proceed. Exiting." && exit 1
-	fi
-	
+
 	"$QUIET" && exec > /dev/null
+	"$QUIET" && exec 2>/dev/null
 	# this is it, deploy the cert.
 	if version_gt "$DETECTED_ZIMBRA_VERSION" "8.7"; then
 		su - zimbra -c "$ZMPATH/bin/zmcertmgr deploycrt comm $tmpcerts/cert.pem $tmpcerts/zimbra_chain.pem -deploy ${SERVICES}"
 	else
-		$ZMPATH/bin/zmcertmgr deploycrt comm "$tmpcerts/cert.pem" "$tmpcerts/zimbra_chain.pem"
+		"$ZMPATH/bin/zmcertmgr" deploycrt comm "$tmpcerts/cert.pem" "$tmpcerts/zimbra_chain.pem"
 	fi
-	"$QUIET" && exec > /dev/tty
+	"$QUIET" && exec > /dev/stdout
+	"$QUIET" && exec 2> /dev/stderr
 
 	! "$QUIET" && echo "Removing temporary files in $tmpcerts"
 	# this is kind of sketchy
 	[ -n "$tmpcerts" ] && rm -r "$tmpcerts"
 
+	set +e
+
 	if "$RESTART_ZIMBRA"; then
-		if ! "$QUIET" && "$PROMPT_CONFIRM"; then
+		if "$PROMPT_CONFIRM"; then
 			prompt "Restart Zimbra?"
-			(( $? == 1 )) && echo "Cannot proceed. Exiting." && exit 1
+			(( $? == 1 )) && echo "Cannot proceed. Exiting." && exit 0
 		fi
 
 		! "$QUIET" && echo "Restarting Zimbra."
 
 		"$QUIET" && exec > /dev/null
+		"$QUIET" && exec 2> /dev/null
 		# Finally apply cert!
-		su - zimbra -c 'zmcontrol restart'
+		su - zimbra -c 'zmcontrol restart' 200>&-
 		# FIXME And hope that everything started fine! :)
-		"$QUIET" && exec > /dev/tty
+		"$QUIET" && exec > /dev/stdout
+		"$QUIET" && exec 2> /dev/stderr
 	fi
-
-	set +e
 
 	return 0
 }
@@ -487,8 +608,7 @@ USAGE: $(basename $0) < -d | -n | -p > [-aNuzjxcq] [-H my.host.name] [-e extra.d
 	 -c | --prompt-confirm: ask for confirmation. Incompatible with -q/--quiet.
 	 -q | --quiet: Do not output on stdout. Useful for scripts. Implies -N/--noninteractive, incompatible with -c/--prompt-confirm.
 
-Author: Lorenzo Milesi <maxxer@yetopen.it>
-Contributors: Jernej Jakob <jernej.jakob@gmail.com> @jjakob
+Authors: Lorenzo Milesi <maxxer@yetopen.it>, Jernej Jakob <jernej.jakob@gmail.com> @jjakob
 Feedback, bugs and PR are welcome on GitHub: https://github.com/yetopen/certbot-zimbra.
 
 Disclaimer:
@@ -606,15 +726,16 @@ done
 
 ## actions
 bootstrap
-get_domain
 
 if ! "$DEPLOY_ONLY"; then
 	if "$NO_NGINX"; then
-		! check_port "$PORT" && echo "Error: port check failed. A web server to use for letsencrypt authentication of the domain $DOMAIN must be listening on the port specified with --port." && exit 1
+		! check_port "$PORT" && echo "Error: port check failed. A web server to use for letsencrypt authentication of the domain $DOMAIN must be listening on the port specified with --port." \
+				&& exit 1
 	else
 		WEBROOT="$ZMWEBROOT"
 		check_zimbra_proxy
-		! check_port "$PORT" nginx zimbra && echo "Error: port check failed. If you have overridden the port with --port, a web server to use for letsencrypt authentication of the domain $DOMAIN must be listening on it." && exit 1
+		! check_port "$PORT" nginx zimbra && echo "Error: port check failed. If you have overridden the port with --port, a web server to use for letsencrypt authentication \
+				of the domain $DOMAIN must be listening on it." && exit 1
 		patch_nginx
 	fi