+++ CHANGELOG.md
@@ -1,3 +1,75 @@
+## v0.7.11
+
+* Fix extracting CA from tls-ca-bundle.pem when spaces are removed from CN (#106)
+    - On some RHEL/CentOS/Fedora systems, tls-ca-bundle.pem lists certificate CN's
+      with spaces removed. This fixes extraction of the CA in such cases.
+
+## v0.7.10
+
+* Add missing check for --prompt-confirm (#100)
+* Improve nginx templates patching process (#101)
+    - replace sed with awk script that is much more resilient,
+      doesn't care about indentation, handles modified templates
+      without breaking
+* Separate nginx templates and includes check
+    - Previously the templates were patched if the includes were unpatched,
+      this happened in the case when the templates were previously patched but
+      the nginx restart was rejected or failed. This resulted in the templates
+      being patched multiple times.
+      Now we check for patched templates first, then check for unpatched includes
+      and restart nginx if necessary.
+* Restore old nginx templates if patching fails, add more status output
+* Rework certbot version detection (#104)
+    - fix improper pipeline syntax
+    - detect certbot errors and grep parsing errors separately
+    - print helpful error message prompting the user to run certbot by itself
+      in case it may need to bootstrap
+* Lowercase local variable
+* Use gawk (GNU awk) and add it to dependency check (fixes #102)
+* Add zmhostname to dependency check
+
+## v0.7.9
+
+* Re-add showing certbot version if prompt-confirm isn't set
+* Move get_domain to bootstrap()
+* Handle interactive certbot bootstrap (#99)
+  - add --no-bootstrap to certbot --version if --prompt-confirm is not set
+      (exit with error if certbot needs to bootstrap)
+  - pass through --noninteractive (answers yes to all certbot questions)
+* README: replace @jjakob e-mail with username
+* README: remove --cert-name from command as it's already passed
+* README: fix typo
+* README: remove reference to 0.5 stable version, tone down warnings
+
+## v0.7.8
+
+* Close fd used for locking on exit, workaround for issue #89
+* Remove lock file too (#89)
+* Tell user he can remove the lockfile
+* Custom ss check, as it doesn't contain the user. Fix #91
+* Add troubleshooting section to readme
+* Add issue template
+* Fix formatting of issue template
+* Add "how it works" explanation
+* Add certbot version check
+* Remove lock file only if we successfully obtained a lock (#89)
+* On quiet redirect to /dev/stdout instead of /dev/tty, also redirect stderr
+* Rewrite find_additional_public_hostnames to be much faster (#95)
+* README: update section describing additional hostname detection
+* Some extra quoting where possible
+* Close locking fd when running zmproxyctl/zmcontrol restart (fix #89)
+* Remove unneeded commented proxy service check
+* Add nonstandard zimbraMailProxyPort warning with prompt (issue #90)
+* Wrap very long lines
+* Update TESTING: moved -e to tested
+* certbot: use --cert-name instead of --expand (issue #97)
+* Change exit status for prompts, fix typos
+* Changed prompt to allow not including detected SANs
+* Add prompt before running certbot
+* Remove redundant QUIET check before prompts
+* Fix prompts exit status in deploy stage
+* Create webroot even without prompt confirm enabled
+
 ## v0.7.7-beta
 
 * fix --extra-domain parsing and additional domain detection (#87)